extern the useMode hook
