+++ EnlightenmentApp.FrontendApp/frontend-react/src/hooks/useMode.tsx
@@ -1,0 +1,21 @@
+import { createContext, useState, useMemo } from "react";
+import { createTheme } from "@mui/material/styles";
+import { PaletteMode } from "@mui/material/index";
+import { themeSettings } from "../theme";
+
+export const ColorModeContext = createContext({
+  toggleColorMode: () => {},
+});
+
+export const useMode = () => {
+  const [mode, setMode] = useState("dark" as PaletteMode);
+
+  const colorMode = useMemo(
+    () => ({
+      toggleColorMode: () => setMode((prev) => (prev === "light" ? "dark" : "light")),
+    }),
+    []
+  );
+  const theme = useMemo(() => createTheme(themeSettings(mode)), [mode]);
+  return [theme, colorMode] as const;
+};
