+++ EnlightenmentApp.FrontendApp/frontend-react/src/theme.tsx
@@ -1,5 +1,4 @@
-import { createContext, useState, useMemo } from "react";
-import { ThemeOptions, createTheme } from "@mui/material/styles";
+import { ThemeOptions } from "@mui/material/styles";
 import { PaletteMode } from "@mui/material/index";
 
 export const tokens = (mode: PaletteMode) => ({
@@ -191,21 +190,4 @@ export const themeSettings = (mode: PaletteMode): ThemeOptions => {
   };
 
   return themeOpts;
-};
-
-export const ColorModeContext = createContext({
-  toggleColorMode: () => {},
-});
-
-export const useMode = () => {
-  const [mode, setMode] = useState("dark" as PaletteMode);
-
-  const colorMode = useMemo(
-    () => ({
-      toggleColorMode: () => setMode((prev) => (prev === "light" ? "dark" : "light")),
-    }),
-    []
-  );
-  const theme = useMemo(() => createTheme(themeSettings(mode)), [mode]);
-  return [theme, colorMode] as const;
 };
